update test to new eager parallel api
