+++ test/auto_parallel/test_semi_auto_parallel_basic.py
@@ -168,7 +168,6 @@ class TestSemiAutoParallelBasic(test_base.CommunicationTestDistBase):
                 user_defined_envs=envs,
             )
 
-
     def test_unary_elementwise_like_api(self):
         envs_list = test_base.gen_product_envs_list(
             self._default_envs, self._changeable_envs
@@ -176,26 +175,6 @@ class TestSemiAutoParallelBasic(test_base.CommunicationTestDistBase):
         for envs in envs_list:
             self.run_test_case(
                 "semi_auto_parallel_for_unary_elementwise_like.py",
-                user_defined_envs=envs,
-            )
-
-    def test_softmax_api(self):
-        envs_list = test_base.gen_product_envs_list(
-            self._default_envs, self._changeable_envs
-        )
-        for envs in envs_list:
-            self.run_test_case(
-                "semi_auto_parallel_for_softmax.py",
-                user_defined_envs=envs,
-            )
-
-    def test_compare_api(self):
-        envs_list = test_base.gen_product_envs_list(
-            self._default_envs, self._changeable_envs
-        )
-        for envs in envs_list:
-            self.run_test_case(
-                "semi_auto_parallel_for_compare.py",
                 user_defined_envs=envs,
             )
