+++ test/cpp/auto_parallel/CMakeLists.txt
@@ -8,11 +8,27 @@ cc_test(
   DEPS proto_desc)
 
 if(WITH_DISTRIBUTE)
+  cc_library(spmd_rule_test_util SRCS spmd_rule_test_util.cc)
   cc_test(
     dist_tensor_test
     SRCS dist_tensor_test.cc
     DEPS phi)
-  cc_test_old(spmd_rule_test SRCS spmd_rule_test.cc DEPS spmd_rules phi)
+  paddle_test(
+    spmd_rule_test
+    SRCS
+    spmd_rule_test.cc
+    DEPS
+    spmd_rule_test_util
+    spmd_rules
+    phi)
+  paddle_test(
+    softmax_grad_spmd_rule_test
+    SRCS
+    softmax_grad_spmd_rule_test.cc
+    DEPS
+    spmd_rule_test_util
+    spmd_rules
+    phi)
 endif()
 
 cc_test_old(dist_mapper_test SRCS dist_mapper_test.cc DEPS phi)
